refactor: move sync button and last sync time to SearchBar sort controls area (#226)
Move the '同步' (Sync) button and its associated last sync timestamp
from the RepositoryList controls bar to the SearchBar component,
positioned to the right of the sort order toggle button (按星标排序 ↓).

- Add sync handler (handleStarSync) and formatLastSync to SearchBar
- Add GitHubApiService, forceSyncToBackend, useDialog imports to SearchBar
- Remove sync button, dead sync code, and unused state from RepositoryList
- Clean up unused imports (RefreshCw) and store destructures (lastSync,
  setRepositories, setLastSync, isSyncing) from RepositoryList
